Split remote fetching-related functions into own file
