+++ tldextract/tldextract.py
@@ -35,7 +35,6 @@ from functools import wraps
 import logging
 import os
 import re
-import socket
 
 import idna
 
@@ -53,6 +52,10 @@ except ImportError:
             path = os.path.join(moddir, resource_name)
             return open(path)
 
+from .remote import find_first_response
+from .remote import looks_like_ip
+from .remote import SCHEME_RE
+
 # pylint: disable=invalid-name,undefined-variable
 try:
     STRING_TYPE = basestring
@@ -60,18 +63,6 @@ except NameError:
     STRING_TYPE = str
 # pylint: enable=invalid-name,undefined-variable
 
-
-# pylint: disable=import-error,invalid-name,no-name-in-module,redefined-builtin
-try:  # pragma: no cover
-    # Python 2
-    from urllib2 import urlopen
-    from urlparse import scheme_chars
-except ImportError:  # pragma: no cover
-    # Python 3
-    from urllib.request import urlopen
-    from urllib.parse import scheme_chars
-    unicode = str
-# pylint: enable=import-error,invalid-name,no-name-in-module,redefined-builtin
 
 LOG = logging.getLogger("tldextract")
 
@@ -82,9 +73,6 @@ PUBLIC_SUFFIX_LIST_URLS = (
     'https://publicsuffix.org/list/public_suffix_list.dat',
     'https://raw.githubusercontent.com/publicsuffix/list/master/public_suffix_list.dat',
 )
-
-SCHEME_RE = re.compile(r'^([' + scheme_chars + ']+:)?//')
-IP_RE = re.compile(r'^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$') # pylint: disable=line-too-long
 
 
 class ExtractResult(collections.namedtuple('ExtractResult', 'subdomain domain suffix')):
@@ -241,7 +229,7 @@ class TLDExtract(object):
 
         tlds = frozenset()
         if self.suffix_list_urls:
-            raw_suffix_list_data = fetch_file(self.suffix_list_urls)
+            raw_suffix_list_data = find_first_response(self.suffix_list_urls)
             tlds = get_tlds_from_raw_suffix_list_data(
                 raw_suffix_list_data,
                 self.include_psl_private_domains
@@ -348,36 +336,6 @@ def get_tlds_from_raw_suffix_list_data(suffix_list_source, include_psl_private_d
     return frozenset(tld_iter)
 
 
-def fetch_file(urls):
-    """ Decode the first successfully fetched URL, from UTF-8 encoding to
-    Python unicode.
-    """
-    text = ''
-
-    for url in urls:
-        try:
-            conn = urlopen(url)
-            text = conn.read()
-        except IOError as ioe:
-            LOG.error('Exception reading Public Suffix List url ' + url + ' - ' + str(ioe) + '.')
-        else:
-            return _decode_utf8(text)
-
-    LOG.error(
-        'No Public Suffix List found. Consider using a mirror or constructing '
-        'your TLDExtract with `suffix_list_url=None`.'
-    )
-    return unicode('')
-
-
-def _decode_utf8(text):
-    """ Decode from utf8 to Python unicode string.
-
-    The suffix list, wherever its origin, should be UTF-8 encoded.
-    """
-    return unicode(text, 'utf-8')
-
-
 class _PublicSuffixListTLDExtractor(object):
 
     def __init__(self, tlds):
@@ -401,21 +359,6 @@ class _PublicSuffixListTLDExtractor(object):
                 return i
 
         return len(lower_spl)
-
-
-def looks_like_ip(maybe_ip):
-    """Does the given str look like an IP address?"""
-    if not maybe_ip[0].isdigit():
-        return False
-
-    try:
-        socket.inet_aton(maybe_ip)
-        return True
-    except (AttributeError, UnicodeError):
-        if IP_RE.match(maybe_ip):
-            return True
-    except socket.error:
-        return False
 
 
 def main():
